Connect to MySQL

+++ src/main/java/cl/edutecno/repository/RatingRepository.java
@@ -6,6 +6,9 @@ import org.springframework.stereotype.Repository;
 import cl.edutecno.model.Rating;
 
 @Repository
-public interface RatingRepository extends JpaRepository<Rating, Long>{
+public interface RatingRepository extends JpaRepository<Rating, Integer>{
+	
+	//@Query("SELECT AVG(r.rating) FROM Rating r WHERE r.show_id = ?1")
+	//Long ratingsPromedio(Long id);
 
 }
