+++ src/main/java/cl/edutecno/repository/ShowRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import cl.edutecno.model.Show;
 
 @Repository
-public interface ShowRepository extends JpaRepository<Show, Long>{
+public interface ShowRepository extends JpaRepository<Show, Integer>{
 
 }
